Fixes for OSX as described by the README.md.

+++ tvision-py/classes/tdesktop.cc
@@ -144,7 +144,7 @@ unsigned iSqr( unsigned i )
 {
     unsigned res1 = 2;
     unsigned res2 = i/res1;
-    while( abs( res1 - res2 ) > 1 )
+    while( abs((long)(res1 - res2) ) > 1 )
         {
         res1 = (res1 + res2)/2;
         res2 = i/res1;
